requirements.txt
flask
Werkzeug
google-cloud-storage
google-cloud-speech
grpcio